scripts/scrapers/requirements.txt
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
kagglehub>=0.1.4
psycopg2-binary>=2.9.0
